+++ changelog.md
@@ -2,6 +2,15 @@
 
 Every new change will be logged
 
+## v1.0.1
+```
++ Updated various Proxy Handlers
++ Updated MailVerifier to wait until the new token is received
++ Updated smooth_out_mouse() to fix https://github.com/Vinyzu/DiscordGenerator/issues/3 by adding an extra point
++ Updated DISCUM Client to reduce code amount and allow proxies to work
+
+```
+
 ## v1.0
 ```
 + First Commit